+++ crates/miden-testing/src/lib.rs
@@ -19,8 +19,6 @@ pub use mock_chain::{
 mod tx_context;
 #[cfg(test)]
 pub(crate) use tx_context::TestTransactionBuilder;
-#[allow(deprecated)]
-pub use tx_context::TransactionContextBuilder;
 pub use tx_context::{ExecError, MockTransaction, MockTransactionBuilder};
 
 pub mod asserts;