+++ crates/miden-testing/src/tx_context/mod.rs
@@ -1,12 +1,9 @@
-mod builder;
 mod context;
 mod errors;
 mod mock_transaction_builder;
 #[cfg(test)]
 mod test_builder;
 
-#[allow(deprecated)]
-pub use builder::TransactionContextBuilder;
 pub use context::MockTransaction;
 pub use errors::ExecError;
 pub use mock_transaction_builder::MockTransactionBuilder;